tests/wpt/tests/workers/WorkerLocation_search.htm

<!DOCTYPE html>
<title> WorkerLocation URL decomposition IDL attribute: search </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/WorkerLocation.js?SearchString');
  worker.onmessage = t.step_func_done(function(e) {
      assert_equals(e.data.search, '?SearchString');
  });
});
</script>
